requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
PyMySQL==1.1.1
PySide6==6.6.3.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0
email-validator==2.1.1
PyInstaller==6.3.0
